+++ cantera18/branches/scons/Cantera/src/thermo/VPSSMgr.cpp
@@ -136,7 +136,7 @@ namespace Cantera {
     // Go see if the SpeciesThermo type is a GeneralSpeciesThermo
     GeneralSpeciesThermo * gst = dynamic_cast<GeneralSpeciesThermo *>(sp_ptr);
     if (gst) {
-      for (int k = 0; k < m_kk; k++) {
+      for (size_t k = 0; k < m_kk; k++) {
 	SpeciesThermoInterpType *st = gst->provideSTIT(k);
 	STITbyPDSS * stpd = dynamic_cast<STITbyPDSS *>(st);
 	if (stpd) {
@@ -193,7 +193,7 @@ namespace Cantera {
     if (m_useTmpStandardStateStorage) {
       std::copy(m_hss_RT.begin(), m_hss_RT.end(), urt);
       doublereal pRT = m_plast / (GasConstant * m_tlast);
-      for (int k = 0; k < m_kk; k++) {
+      for (size_t k = 0; k < m_kk; k++) {
 	urt[k] -= pRT * m_Vss[k];
       }
     } else {
@@ -311,7 +311,7 @@ namespace Cantera {
   }
 
   void VPSSMgr::_updateStandardStateThermo() {
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
       PDSS *kPDSS = m_vptp_ptr->providePDSS(k);
       kPDSS->setState_TP(m_tlast, m_plast);
     }
@@ -321,7 +321,7 @@ namespace Cantera {
   void VPSSMgr::_updateRefStateThermo() const {
     if (m_spthermo) {
       m_spthermo->update(m_tlast, &m_cp0_R[0], &m_h0_RT[0], &m_s0_R[0]);
-      for (int k = 0; k < m_kk; k++) {
+      for (size_t k = 0; k < m_kk; k++) {
 	m_g0_RT[k] = m_h0_RT[k] - m_s0_R[k];
       }
     }
@@ -367,7 +367,7 @@ namespace Cantera {
   void VPSSMgr::initThermoXML(XML_Node& phaseNode, std::string id) {
     const PDSS *kPDSS = m_vptp_ptr->providePDSS(0);
     m_p0 = kPDSS->refPressure();
-    for (int i = 0; i < m_kk; i++) {
+    for (size_t i = 0; i < m_kk; i++) {
       const PDSS *kPDSS = m_vptp_ptr->providePDSS(i);
       doublereal mint = kPDSS->minTemp();
       if (mint > m_minTemp) {
@@ -382,7 +382,7 @@ namespace Cantera {
     // Add a check to see that all references pressures are the same
     double m_p0_k;
     if (m_spthermo) {
-      for (int k = 0; k < m_kk; k++) {
+      for (size_t k = 0; k < m_kk; k++) {
 	m_p0_k = m_spthermo->refPressure(k);
 	if (m_p0 != m_p0_k) {
 	  //throw CanteraError("VPSSMgr::initThermoXML",
@@ -395,7 +395,7 @@ namespace Cantera {
       }
     }
 
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
       const PDSS *kPDSS = m_vptp_ptr->providePDSS(k);
       m_p0_k = kPDSS->refPressure();
       if (m_p0 != m_p0_k) {
